Removed 'assert' and 'expect' macros.
The framework is changing. Needs from the nuvm project showed that the
assertion/expecting of values is better done on the user's side. It may
yet be that the framework will incorporate the macros used on the nuvm
project and standardize around these, but for now this project is merely
a support project for nuvm.

Such as it is, I'll be experimenting with the possibilities on nuvm's
side and only port them back here when they're a little more stable.